+++ example/lib/presentation/samples/bar/bar_chart_sample8.dart
@@ -1,0 +1,145 @@
+import 'dart:math';
+
+import 'package:fl_chart_app/presentation/resources/app_resources.dart';
+import 'package:fl_chart_app/util/extensions/color_extensions.dart';
+import 'package:fl_chart/fl_chart.dart';
+import 'package:flutter/material.dart';
+
+class BarChartSample8 extends StatefulWidget {
+  BarChartSample8({super.key});
+
+  final Color barBackgroundColor =
+      AppColors.contentColorWhite.darken().withOpacity(0.3);
+  final Color barColor = AppColors.contentColorWhite;
+
+  @override
+  State<StatefulWidget> createState() => BarChartSample1State();
+}
+
+class BarChartSample1State extends State<BarChartSample8> {
+  @override
+  Widget build(BuildContext context) {
+    return AspectRatio(
+      aspectRatio: 1,
+      child: Padding(
+        padding: const EdgeInsets.all(16),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+          children: <Widget>[
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                const Icon(Icons.graphic_eq),
+                const SizedBox(
+                  width: 32,
+                ),
+                Text(
+                  'Sales forecasting chart',
+                  style: TextStyle(
+                    color: widget.barColor,
+                    fontSize: 18,
+                    fontWeight: FontWeight.bold,
+                  ),
+                ),
+              ],
+            ),
+            const SizedBox(
+              height: 32,
+            ),
+            Expanded(
+              child: BarChart(
+                randomData(),
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  BarChartGroupData makeGroupData(
+    int x,
+    double y,
+  ) {
+    return BarChartGroupData(
+      x: x,
+      barRods: [
+        BarChartRodData(
+          toY: y,
+          color: x >= 4 ? Colors.transparent : widget.barColor,
+          borderRadius: BorderRadius.zero,
+          borderDashArray: x >= 4 ? [4, 4] : null,
+          width: 22,
+          borderSide: BorderSide(color: widget.barColor, width: 2.0),
+        ),
+      ],
+    );
+  }
+
+  Widget getTitles(double value, TitleMeta meta) {
+    const style = TextStyle(
+      color: Colors.white,
+      fontWeight: FontWeight.bold,
+      fontSize: 14,
+    );
+    List<String> days = ['M', 'T', 'W', 'T', 'F', 'S', 'S'];
+
+    Widget text = Text(
+      days[value.toInt()],
+      style: style,
+    );
+
+    return SideTitleWidget(
+      axisSide: meta.axisSide,
+      space: 16,
+      child: text,
+    );
+  }
+
+  BarChartData randomData() {
+    return BarChartData(
+      maxY: 300.0,
+      barTouchData: BarTouchData(
+        enabled: false,
+      ),
+      titlesData: FlTitlesData(
+        show: true,
+        bottomTitles: AxisTitles(
+          sideTitles: SideTitles(
+            showTitles: true,
+            getTitlesWidget: getTitles,
+            reservedSize: 38,
+          ),
+        ),
+        leftTitles: const AxisTitles(
+          sideTitles: SideTitles(
+            reservedSize: 30,
+            showTitles: true,
+          ),
+        ),
+        topTitles: const AxisTitles(
+          sideTitles: SideTitles(
+            showTitles: false,
+          ),
+        ),
+        rightTitles: const AxisTitles(
+          sideTitles: SideTitles(
+            showTitles: false,
+          ),
+        ),
+      ),
+      borderData: FlBorderData(
+        show: false,
+      ),
+      barGroups: List.generate(
+        7,
+        (i) => makeGroupData(
+          i,
+          Random().nextInt(290).toDouble() + 10,
+        ),
+      ),
+      gridData: const FlGridData(show: false),
+    );
+  }
+}